Added an Auto Run feature to the config file
Specify the story you want to auto run, then it will automatically open to the autosave of that story (or if an autosave is not present than it will start a new game). If Autorun is not specified in the config file then it takes you to the main menue instead.